Tweak behavior when attacking an empty enemy-occupied cell.
That is, the cell is notionally "occupied" by another player, but the
amount of goop there is 0. Previously this would result in the cell
being occupied by the new invading player, but still with 0 goop.
Clearly that should be 1 goop. So change the reference implementation to
match the new test.